기능: PlayerSettingsSnapshot 구조체 분리 (B #1/8)
기존 AITPlayerSettingsBackup 의 필드와 Capture/Restore 를 PlayerSettingsSnapshot
으로 이전. [Serializable] 부여로 향후 ScriptableSingleton 직렬화 대응. 기존
AITPlayerSettingsBackup 은 [Obsolete] 얇은 래퍼로 유지 — 호출부 교체는 후속 Task.
