+++ Editor/AITBuildSession.cs.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 8b2c4e9a1d7f4c5d9a2e3f6b7c8d9e0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  references: []
