+++ Tests~/E2E/SharedScripts/Editor/EditModeTests/AITBuildSessionTests.cs.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 9c3d5fab2e8f5d6eab3f4a7c8d9e0f1b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  references: []
